refactor: create reusable `BlockGenContext` from `FunctionContext` basics
